move generate_signature and verify_signature from images/views/serve to images/utils
