+++ wagtail/images/tests/tests.py
@@ -14,7 +14,8 @@ from wagtail.images.formats import Format, get_image_format, register_image_form
 from wagtail.images.forms import get_image_form
 from wagtail.images.models import Image as WagtailImage
 from wagtail.images.rect import Rect, Vector
-from wagtail.images.views.serve import ServeView, generate_signature, verify_signature
+from wagtail.images.utils import generate_signature, verify_signature
+from wagtail.images.views.serve import ServeView
 from wagtail.tests.testapp.models import CustomImage, CustomImageFilePath
 from wagtail.tests.utils import WagtailTestUtils
